requirements.txt
discord.py
flake8
wget
requests
boto3
tinydb